Servicios proceso compra 1

+++ models/rutapunto.js
@@ -3,8 +3,8 @@ module.exports = function(sequelize, DataTypes) {
   var RutaPunto = sequelize.define('RutaPunto', {
     indice: DataTypes.INTEGER,
     descripcion: DataTypes.STRING,
-    latitud: {type: DataTypes.FLOAT, allowNull: false, validate: { min: -90, max: 90 } },
-    longitud: {type: DataTypes.FLOAT, allowNull: false, validate: { min: -180, max: 180 } },
+    latitud: {type: DataTypes.FLOAT(14,10), allowNull: false, validate: { min: -90, max: 90 } },
+    longitud: {type: DataTypes.FLOAT(14,10), allowNull: false, validate: { min: -180, max: 180 } },
     tipo: DataTypes.INTEGER
   }, {
     timestamps: false,
